Ensure that JolokiaMvcEndpoint destroys underlying controller
Closes gh-7606
